requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
Django==4.2.2
django-seo-plus==0.0.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
